requirements.txt
aiohttp
aiosqlite
discord.py==2.6.3
python-dotenv
discord-interactions
requests
